dev/llm_service/requirements.txt
aio-pika>=9.0.0
asyncpg>=0.29.0
httpx>=0.25.0
python-dotenv>=1.0.0
